requirements.txt
# Core Flask dependencies
flask>=2.3.0
python-dotenv>=1.0.0
werkzeug>=2.3.0

# Authentication and OAuth
authlib>=1.2.0
requests>=2.31.0
bcrypt>=4.0.0

# PDF and QR code generation
reportlab>=4.0.0
qrcode[pil]>=7.4.0
pillow>=10.0.0

# Email functionality
validators>=0.22.0
bleach>=6.0.0

# Production server
gunicorn>=21.0.0

# Enhanced features (optional)
APScheduler>=3.10.0
flask-limiter>=3.5.0

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-flask>=1.2.0
# coverage>=7.3.0
